fix: 빌드 에러로 인한 클라이언트 컴포넌트 분리
