+++ src/app/people/layout.tsx
@@ -1,38 +1,15 @@
-'use client'
+import { Suspense } from 'react'
 
-import { useEffect, useState } from 'react'
-
-import { useRouter, useSearchParams } from 'next/navigation'
-
-import SideMenuLayout from '@src/components/SideMenuLayout'
-import { PEOPLE_NAVIGATION } from '@src/constants/people'
+import PeopleLayoutContent from '@src/containers/people/LayoutContent'
 
 export default function PeopleLayout({
   children,
 }: Readonly<{
   children: React.ReactNode
 }>) {
-  const router = useRouter()
-  const searchParams = useSearchParams()
-  const [activeSection, setActiveSection] = useState(PEOPLE_NAVIGATION[0])
-
-  useEffect(() => {
-    const type = searchParams.get('type') ?? PEOPLE_NAVIGATION[0]
-    setActiveSection(type.toUpperCase())
-  }, [searchParams])
-
-  const onChangeSection = (menu: string) => {
-    setActiveSection(menu)
-    router.push(`/people?type=${menu}`)
-  }
-
   return (
-    <SideMenuLayout
-      sectionList={PEOPLE_NAVIGATION}
-      activeSection={activeSection}
-      onChangeSection={onChangeSection}
-    >
-      {children}
-    </SideMenuLayout>
+    <Suspense fallback={<p>Loading...</p>}>
+      <PeopleLayoutContent>{children}</PeopleLayoutContent>
+    </Suspense>
   )
 }
